Reorganized entire website to allow for routing through App.tsx file, new pages folder holds all pages
