+++ package.json
@@ -13,9 +13,10 @@
     "bootstrap": "^5.3.2",
     "gh-pages": "^6.1.1",
     "openai": "^4.26.0",
-    "react": "^18.2.0",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.10.1",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^7.4.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
