+++ src/pages/BasicQuiz.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export function BasicQuiz(): React.JSX.Element {
+    return <div>
+        Basic Quiz
+    </div>
+}
