+++ src/pages/DetailedQuiz.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export function DetailedQuiz(): React.JSX.Element {
+    return <div>
+        Detailed Quiz
+    </div>
+}
